requirements.txt
# FMO Echo 服务依赖项
#
# 安装方法:
#   pip install -r requirements.txt

# MQTT 客户端库
paho-mqtt>=1.6.0

# YAML 配置文件解析
PyYAML>=6.0
